+++ app/src/main/java/com/anonymous/appilogue/features/home/bottomsheet/hole/BottomSheetHoleAppAdapter.kt
@@ -1,4 +1,4 @@
-package com.anonymous.appilogue.features.home.bottomsheet
+package com.anonymous.appilogue.features.home.bottomsheet.hole
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -9,10 +9,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.anonymous.appilogue.R
 import com.anonymous.appilogue.databinding.ItemAppBinding
 import com.anonymous.appilogue.model.Review
-import com.anonymous.appilogue.model.ReviewedApp
 
-class BottomSheetAppAdapter :
-    ListAdapter<Review, BottomSheetAppAdapter.BottomSheetAppViewHolder>(diffCallback) {
+class BottomSheetHoleAppAdapter :
+    ListAdapter<Review, BottomSheetHoleAppAdapter.BottomSheetAppViewHolder>(diffCallback) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BottomSheetAppViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
