+++ app/src/main/java/com/anonymous/appilogue/features/home/bottomsheet/planet/MyPlanetFragment.kt
@@ -1,7 +1,5 @@
-package com.anonymous.appilogue.features.home.bottomsheet
+package com.anonymous.appilogue.features.home.bottomsheet.planet
 
-import android.os.Bundle
-import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.anonymous.appilogue.R
