+++ app/src/main/res/layout/fragment_home.xml
@@ -17,7 +17,7 @@
 
         <variable
             name="mySpaceDustViewModel"
-            type="com.anonymous.appilogue.features.home.bottomsheet.MySpaceDustViewModel" />
+            type="com.anonymous.appilogue.features.home.bottomsheet.space_dust.MySpaceDustViewModel" />
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
